Removed order dependency in named graph test.
The TestNamedGraphs class assumed that the names of the graphs in the dataset would
appear in the listNames iterator in the same order as they were added to the dataset.
However this is not and should not be part of the contract, and at least one implementation
stores the names in a set, hence scrambling the order.

Fix is to make the test order-independant, checking the models by name, checking
that the returned names are the same as the given names of the graphs.