Fixed ordering problem within XML Gold tests by imposing a specific order.
Two of the thirteen XML gold tests were failing under Java 8 because the order that items were generated
in differed from that in the saved gold files under elda-lda/src/test/resources/xml_gold/*. The orders
differed for format-names and view-names because they were fulled from sets of those names and hence
the order is unspecified.

We replaced the set of names by a list of names and sorted that list in the default order so
that it would be the same every time. Then the xml_rendering for the failing test cases
was updated with the updated rendering.

A note in passing: sorting out the golds is currently too complicated, needing lots of manual
intervention.