Resolved ordering bug in on-the-fly shortname bindings in JSON renderer encoding.
An long-standing (and probably obsolete) round-trip test testNameClash1 failed when Elda was
tested under Java 8. This was traced down to the code that produced a sorted list of
properties for a given resource. The sorting is done on the shortname of the property,
and the shortnames are potentially allocated by the comparator during the sort. This
means that the order the shortnames are allocated in is unspecified and may change.
In the presenting problem, a URI with localname 'type' was competing with 'rdf:type'
and could (did) claim the shortname 'type', which was unexpected, so the test failed.

Further analysis revealed that when other:type was bound to the shortname 'type',
and rdf:type was consequently bound to the shortname rdf_type, the serialisation
still used 'type' as rdf:types shortname. This was tracked down to ContextPropertyInfo's
getSerialisationName, which used the local name of the given URI if it was in any of
the namespaces rdf, rdfs, api, or khv. This looks like an old hack for avoiding
user-defined URIs clashing with terms used by Elda for rendering control if they
happen to have the same local name and hash been superseded by better construction
of shortname tables. Suppressing this hack does not cause any of the Elda tests
to fail and so it stays suppressed.

testNameClash1 was modified to give rdf:type a context property entry before running
the Encoder so that whatever order the properties are allocated in while rendering
the 'user' property does not override the 'system' property. The Encoder was modified
so that for any given resource, its properties have their context info constructed
in alphabetical order of property URI.

It's very heavy to do all this sorting and property lookup for each resource in
the rendering. It needs a rethink.