Revision of labelOf to fix ordering.
In XMLRendering the objects of properties are renderered in a
fixed order determined by their 'label'. If there is no explicit
label to use, the local name of a resource is used. Sadly the
'local name' of (eg) 'whatever:pathbit.xml?_metadata=all' is 'all',
as it is for .rdf and .ttl etc. This makes the ordering
unspecified between .xml and .rdf etc, and this shows up in
differences between Java 6/7 and Java 8.

A possible fix is to strip away the quary parameters. This
code is conditionally included while we're trying it out.