Saved gold files for xml rendering tests.
Some tinkering with the code in XMLRenderer and TestXMLRendererWithGold
to make it easier to change the gold files -- made an explicit boolean
flag in XMLRenderer for the gold saving (which ought not to be in non-
test code but I can't see how to do better right now) and ditto in
TestXMLRendererWithGold which, if enabled, will copy the latest rendered
XML from /tmp to the appropriate gold file for the current test.

Passes under Java 8, awaiting tests under Java 7.