Use entire URI if no short form for ordering objects in XML rendering.
Rather than using an ad-hoc short for, use the URI for labelOf if
no short form has been allocated. The labelOf sorting is only
meant to be repeatable and to coincide with label order if
possible, so using the full URI to sort by is harmless and
gives a well-defined order.