Correct the implementation of the /meta qualifier in URIs.
See issue #131.

There were two problems. The wrong router was being used to
look up the URI tail (that part after the /meta), the
DefaultRoutyer rather than the Router for this servlet.
Fixed by passing in the ServletContext to the lookup
method in ControlRestlet.

Secondly the config display code was not taking account of
the API Specs possible prefix, so the URI path was being
compared against the wrong string. Corrected by handling
the prefix if any is present.