Renamed NameUtils classes.
There were two classes called NameUtils in Elda lda, one in
the main and the other in test. This got in the way when
adding LogRequestFilter which referred to yet another
NameUtils class, this time from lib. Eclipse would not import
the correct package.

Renamed the lda classes to two different names, EldaNameUtils
and NameSupport.