api-config no longer crashes after rendering an endpoint.
1.3.9 introduced a misfeature whereby /api-config crashed with
a NullPointerException once an endpoint had been rendered.

This was due to a new feature exposing an existing bug:

* the variables named _servletResponse and _servletRequest are bound to the
  request's servlet request and response as raw Java values. Calling
  Bindings.get() on those names will return null as the values are not
  of type Value or String.

* when RouterRestlet fetches the default bindings from the API Endpoint
  Spec, it did not make a copy of the Bindings before injecting the
  servlet parameters.

Hence the bindings of _servletRequest/Response persist in the endpoint
spec. /api-config (which is not invoked via RouterRestlet) sees these
bindings and attemts to display them with code that does not expect
null values and so crashes.

Similarly the values of _view, _properties, _page, and _transaction
are pushed into the shared Bindings. This would cause serious
problems if multiple Elda threads concurrently render the same
endpoint (serious in that the values of those variables become
at best unpredictable; they are meant for renderers).

Fix: .copy() the Bindings taken from the endpoint spec.

Also as an advance precaution modify ComposeConfigDisplay so
that if the value of an LDA variable is null it is not displayed
but merely generates a debug message.

Also fixed some spelling and layout mistakes in Bindings and
arranged for all constructors to funnel through the three-argument
constructor.