Make request cache key a "canonical" URI.
See issue #144.

New method getCanonicalURI() returns the request URI
with the renderer name presented as trailing .suffix
of the path and any _format= query parameter removed.
This means that renderers selected by content negotiation
are visible in the key.

(We could in principle canonise the rest of the URI but
that hasn't shown up as a problem yet.)