Mostly-cloning constructor of APIResultSet.
Added a new constructor for APIResultSet that copies an existing
result set except that it has a new MergedModel with the same
object data but no metadata.

This is a step toward reducing the cache footprint of a result
set at thje expense of recomputing the metadata. The same
object data can be used for different renderings. The amount
of metadata is moderately small.