Collapsed cache keys that would generate the same object model.
Now we generate the metadata model afresh whether we create a new
response or retrieve on from the cache, we can use the same cache
entry for all request URIs that would generate the same result
list and model. So we canonise the request uri using

    getURIplain()

which discards _format, _callback, _mark, and _metadata, none
of which influences the retrieved result.