Improving the use (or not) of BytesOutTimed.
Renderers should render completely before writing the response,
because the rendering might fail and we don't want to have already
written 200 OK before the response explodes.

So we define that Renderer.render should compute the reponse
body and retain it briefly before BytesOut.writeAll sends it
on its way. We assume that nothing is going to break as we
send the string content.

So we now have a StreamUtils for sending a content string as
UTF8 down a supplied stream, and we're worjking on using this
in the renderers, especially the Feed and RDF/XML renderers which
currently defer rendering to the writeAll stage.