Implement poisoning of broken responses.
Addresses issue #46.

We would like to (a) allow streaming where possible in Elda
renderers and (b) handle problems when a renderer fails while
the outut is streaming.

Previously the preferred design choice has been to pre-render
the result into a string and then when the response is committed
to send the string. It is unlikely (but not impossible) that
sending the string will fail and so at the point of committing
to output either (1) an exception has been thrown and the error
pages take control or (2) we cann send a 200 OK and reasonably
exopected that the generated string will make it completely
to the end.

However that choice prohibits (locally) streaming output, since
the rendered string must be saved somewhere. (There are many places
where Elda freely builds up collections and does something with
them rather than doing the something as the collection
components fly past.) So we decide to permit streaming output
with computation but arrange that if an exception is thrown
while the computation streams that the output stream is
"poisioned" by adding to it bytes that will (should) cause
the client to fail, eg when reading XML to send <SOMTHINGILLEGAL
(complicated because we don't know what the syntactic
context is).

RouterRestlet (which is where incoming queries are handled)
accepts a BytesOut object which contains (a promise of) the
output stream and turns it into a JAX-RS StreamingOutput
object. When that object is finally provoked into sending output,
the RouteRestlet.wrap method traps any exception and
sends a response-specific poison string before logging the
problem and returning.