Fix failure to log term bindings required by metadata.
When termBindings metadata is generated, terms that do not
appear in the model are dropped to reduce overheads
(both mechanical and cognitive). However this does
not account for terms in the meta-model (eg versions)
and so Elda was generating result models where the
not all shortname -> uri transforms were present.

This mattered when the property was RDFS.label because
the XSLT renderer used thje word `label` to dig items
(such as the format-choosing list on the HTML page)
from the metadata. As a result it would generate strange
tooltips and hrefs.

To fix this, we ensure that the metadata modelettes are
included in the "present in the model" test performed
when generating termbindings.