Changes to handle spaces (and other escapable characters) in URIs.
Addresses issue #156.

The @PathParam that Elda uses to match against URI templates
is the decoded path fragment of the request URI of the request
being processed. Because it is decoded it may contain literal
spaces. Consequently URIs constructed from those path pieces
may also contain spaces. To avoid the generated SPARQL query
causing a SPARQL lexical error, hence an exception, hence a
termination with status 500, hence failure of the query and
any others on the request chain, we wish to re-encode the
space.

(The argument applies more generally but space is the specific
case that has caight us out)

When those URIs are constructed any spaces in them are replaced
by %20 in URIUtils.escapeAsURI. A more general replacement
(from a library rather than one we write ourselves!) is
desired.