EXPERIMENTAL exploration of using inverse properties. Use inv_ as prefix to shortname denoting the inverse property of the remaining shortname.
When constructing a Param.Info spot use of inv_ and give the
Info an inverse value of true.

When APIQuery wants to create a triple for the basic graph
pattern, given its Param.Info I, it passes the S and O values
for the triple to I which uses the value on inverse to decide
whether the triple is constructed forwards (the usual case)
or backwards (the inverse case).