Revisions of license display code.
The ResultSet returns a set of Resource rather than a set of String.
The Resources have properties from the data or config that can be
used to control the layout of a liscense object.

For the moment the expansion of the desription is done in the VelocityRending
class but it will be moved to a velocity macro file.