EndpointMetadata.addTermBindings did not take the model into account when constructing the term bindings, so unknown predicates did not generate bindings. The test
    TestGeneratedMetadata.testTermBindingsCoverAllPredicates

which was supposed to cover this case was incorrect -- it added
the model to the context before doing the termbindings, but the
non-test code did not.

Removed the incorrect inclusion of the model to the context
from the test caused a test failure which was then fixed by
adding the model in addTermBindings.