ValTranslator now checks URIs for SPARQL compatability.
When ValTranslator converts a query parameter to a RDFQ URI object,
it now checks that the URI string is acceptable to SPARQL as
a URI body. This we do by the sledgehammer approach of
constructing a SELECT query

	SELECT (<GIVEN-URI> AS ?x) WHERE {}"

If this throws a QueryParseException then we throw a BadRequest
exception, which in turn will evoke a status 400 with a suitable
error message mentioning the bad URI.

In the unlikely event that this check needs to be switched
off, define the config variable _checkIRISyntax to "no".

Fixes #183.